+++ CHANGELOG.md
@@ -13,6 +13,7 @@ SPDX-License-Identifier: Apache-2.0
 
 - Added Pauli I for Hamiltonian construction.
 - Added Reconfigurable Beam Splitter (RBS) gate.
+- Added the Remote class.
 
 ## [0.8.0](https://pypi.org/project/ket-lang/0.8.0)